requirements.txt

streamlit==1.35.0
scikit-learn==1.5.0
Pillow==10.3.0
PyPDF2==3.0.1
pdfplumber==0.11.0
spacy==3.7.4
nltk==3.8.1
